Update LogitLib.jl
The original Logit dgp always generates samples with close to 50% 1, 50% zero. I believe that this is the reason that training sometimes gets stuck. This modification increases the dispersion, for some parameter draws, and in my trials,  training is not getting stuck.

If used, we need to update the Logit  MLE results, too. I can do that, if you like this PR.